Move file & fix paths
