+++ proxies/collection/schemas/index.js
@@ -2,7 +2,7 @@ import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import _ from 'lodash';
 import { Proxies } from '../';
 
-import registeredProxies from './registered_proxies';
+import registeredProxies from '../registered_proxies';
 
 import apiUmbrellaSchema from './api_umbrella';
 import emqttSchema from './emqtt';
